% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonomy.ch-package.R
\docType{package}
\name{autonomy.ch-package}
\alias{autonomy.ch}
\alias{autonomy.ch-package}
\title{autonomy.ch: Build A Package From Rmarkdown File}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\author{
\strong{Maintainer}: Christine Hoogland \email{christine.hoogland@finddx.org} (\href{https://orcid.org/0000-0002-7341-1551}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
